I made a bullet.h, an abstract class, and a machineGunBullet.cpp, implementing the class and the standard bullet that’ll be shot for this prototype.
Constructing machineGunBullet requires the location of the two ships and the location of the turret. I’m guessing the turret class that’s being worked on will ideally use the class as an attribute.

For full functionality, we would need to dynamically create bullets with a list or something. This way, we can keep track of what bullets to draw. They should, while moving, check to see if they are out of bounds similarly to the rocket class and destruct if necessary.

Without the turret class, I won’t be able to do much more, so I’ll go ahead and push the code I have.